First pass on the bft actor.
