+++ node/actors/bft/src/replica/mod.rs
@@ -4,12 +4,11 @@
 
 mod block;
 pub(crate) mod leader_commit;
-pub(crate) mod leader_prepare;
 mod new_view;
+pub(crate) mod proposal;
 pub(crate) mod replica_prepare;
 mod state_machine;
 #[cfg(test)]
 mod tests;
-mod timer;
 
 pub(crate) use self::state_machine::StateMachine;
